requirements.txt
Django==3.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-mysql==3.8.1
djangorestframework==3.12.1

gunicorn==20.0.4

mysqlclient==2.0.1
PyJWT
python-dotenv==0.14.0
djangorestframework-simplejwt